Changed variable type for path_hint

+++ include/cista/containers/rtree.h
@@ -631,7 +631,7 @@ struct rtree {
       return;
     }
 
-    int h = 0;
+    size_t h = 0;
     h = path_hint_[depth];
     rect crect;
     if (h < delete_node.count_) {
